- Added setting to configure a delay before testing internet connectivity after a wifi connection was established. - Updated osmdroid to 4.0 - Changed map tile provider from Mapnik to MapquestOSM to comply with OSM usage policy. - Removed some notification icon resolutions that don't really seem necessary, looks better now on 4.2.2.

+++ inetify/inetify-test/src/net/luniks/android/inetify/test/TestTester.java
@@ -50,7 +50,7 @@ public class TestTester implements Tester {
 		return info;
 	}
 
-	public TestInfo testWifi(int retries, long delay) {
+	public TestInfo testWifi(int retries, int delay) {
 		done.set(false);
 		testCount.incrementAndGet();
 		cancelled.set(false);
